Move annotation impl util out
