+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTAnnotation.java
@@ -5,11 +5,7 @@
 
 package net.sourceforge.pmd.lang.java.ast;
 
-import java.util.Arrays;
-import java.util.List;
-
 import net.sourceforge.pmd.Rule;
-import net.sourceforge.pmd.lang.java.typeresolution.TypeHelper;
 
 
 /**
@@ -60,41 +56,7 @@ public interface ASTAnnotation extends TypeNode, ASTMemberValue {
      */
     // @formatter:on
     default boolean suppresses(Rule rule) {
-        return AnnotationImplUtil.suppresses(this, rule);
+        return AstImplUtil.suppresses(this, rule);
     }
 }
 
-final class AnnotationImplUtil {
-
-    private static final List<String> UNUSED_RULES
-        = Arrays.asList("UnusedPrivateField", "UnusedLocalVariable", "UnusedPrivateMethod", "UnusedFormalParameter");
-
-    private static final List<String> SERIAL_RULES = Arrays.asList("BeanMembersShouldSerialize", "MissingSerialVersionUID");
-
-    private AnnotationImplUtil() {
-
-    }
-
-    private static boolean isSuppressWarnings(ASTAnnotation astAnnotation) {
-        return TypeHelper.isA(astAnnotation, "java.lang.SuppressWarnings");
-    }
-
-    public static boolean suppresses(ASTAnnotation annotation, Rule rule) {
-        // if (SuppressWarnings.class.equals(getType())) { // typeres is not always on
-        if (isSuppressWarnings(annotation)) {
-            for (ASTLiteral element : annotation.findDescendantsOfType(ASTLiteral.class)) {
-                if (element.hasImageEqualTo("\"PMD\"") || element.hasImageEqualTo(
-                    "\"PMD." + rule.getName() + "\"")
-                    // Check for standard annotations values
-                    || element.hasImageEqualTo("\"all\"")
-                    || element.hasImageEqualTo("\"serial\"") && SERIAL_RULES.contains(rule.getName())
-                    || element.hasImageEqualTo("\"unused\"") && UNUSED_RULES.contains(rule.getName())
-                    || element.hasImageEqualTo("\"all\"")) {
-                    return true;
-                }
-            }
-        }
-
-        return false;
-    }
-}
